refactor(spark): pull shared HoodieSchemaUtils bodies into family base classes
Nothing in this family is identical across all six Spark versions, so the
bodies are shared at the Spark-major-version level: HoodieSpark3SchemaUtils in
hudi-spark3-common carries getSchema (pre-4.0 JdbcUtils.getSchema signature
without the Connection argument), and HoodieSpark4SchemaUtils in
hudi-spark4-common carries all three methods, which are uniform across
4.0-4.2, turning those objects into empty declarations. Spark 3.x objects
keep the per-version deltas: the 3-arg SchemaUtils.checkColumnNameDuplication
on 3.3 (colType parameter removed in Spark 3.4) and StructType#toAttributes
on 3.3/3.4 (moved to DataTypeUtils in Spark 3.5, SPARK-44353).
