+++ hudi-spark-datasource/hudi-spark3.4.x/src/main/scala/org/apache/spark/sql/HoodieSpark34SchemaUtils.scala
@@ -20,17 +20,13 @@
 package org.apache.spark.sql
 
 import org.apache.spark.sql.catalyst.expressions.Attribute
-import org.apache.spark.sql.execution.datasources.jdbc.JdbcUtils
-import org.apache.spark.sql.jdbc.JdbcDialect
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.SchemaUtils
-
-import java.sql.{Connection, ResultSet}
 
 /**
  * Utils on schema for Spark 3.4.
  */
-object HoodieSpark34SchemaUtils extends HoodieSchemaUtils {
+object HoodieSpark34SchemaUtils extends HoodieSpark3SchemaUtils {
   override def checkColumnNameDuplication(columnNames: Seq[String],
                                           colType: String,
                                           caseSensitiveAnalysis: Boolean): Unit = {
@@ -39,13 +35,5 @@ object HoodieSpark34SchemaUtils extends HoodieSchemaUtils {
 
   override def toAttributes(struct: StructType): Seq[Attribute] = {
     struct.toAttributes
-  }
-
-  override def getSchema(conn: Connection,
-                         resultSet: ResultSet,
-                         dialect: JdbcDialect,
-                         alwaysNullable: Boolean = false,
-                         isTimestampNTZ: Boolean = false): StructType = {
-    JdbcUtils.getSchema(resultSet, dialect, alwaysNullable)
   }
 }
