+++ hudi-spark-datasource/hudi-spark4.2.x/src/main/scala/org/apache/spark/sql/HoodieSpark42SchemaUtils.scala
@@ -19,34 +19,7 @@
 
 package org.apache.spark.sql
 
-import org.apache.spark.sql.catalyst.expressions.Attribute
-import org.apache.spark.sql.catalyst.types.DataTypeUtils
-import org.apache.spark.sql.execution.datasources.jdbc.JdbcUtils
-import org.apache.spark.sql.jdbc.JdbcDialect
-import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.util.SchemaUtils
-
-import java.sql.{Connection, ResultSet}
-
 /**
- * Utils on schema for Spark 3.4+.
+ * Utils on schema for Spark 4.2.
  */
-object HoodieSpark42SchemaUtils extends HoodieSchemaUtils {
-  override def checkColumnNameDuplication(columnNames: Seq[String],
-                                          colType: String,
-                                          caseSensitiveAnalysis: Boolean): Unit = {
-    SchemaUtils.checkColumnNameDuplication(columnNames, caseSensitiveAnalysis)
-  }
-
-  override def toAttributes(struct: StructType): Seq[Attribute] = {
-    DataTypeUtils.toAttributes(struct)
-  }
-
-  override def getSchema(conn: Connection,
-                         resultSet: ResultSet,
-                         dialect: JdbcDialect,
-                         alwaysNullable: Boolean = false,
-                         isTimestampNTZ: Boolean = false): StructType = {
-    JdbcUtils.getSchema(conn, resultSet, dialect, alwaysNullable)
-  }
-}
+object HoodieSpark42SchemaUtils extends HoodieSpark4SchemaUtils
